Add documents.index route with overview of documents and revisions
The documents.index route lists all documents, their revisions (up to
what the backend supplies) and allows you to go to the edit view of
these documents.  It is not styled at all, but the overarching
documents route may become the hook under which we place the signing
actions.

The actual contents of the index route may change still.