Make publish flow work on document-container
The publish flow tried to work on an editor-document.  Except it
didn't.  Because it didn't strange errors occurred.  This has been
altered.  The signing flow now starts from the editor-document and
allows singing of the latest document.  In the future this could be
changed to a predefined document which was prepared for this use.