Basics for letting an agenda be published
The reloading is terribly unclean.  We should have seen this was going
to be bad when we resorted to the RSVP.hash in the model.  Other code
seems to be in a fairly clean state without too much state that's hard
to understand.  Most problematic is code that's still lingering
around, or state which the design represents that isn't implemented
yet.