callback endpoint for mow
this is an example endpoint as defined in GN-270.
It can be extended later on to redirect to the correct document etc.

Note that any queryParams that need to be remembered need to be defined
in one of the routes. The mock logic defined in fetch-rdfa is an extension of the
AuthenticatedRouteMixin and is depended on its behaviour.