participants of a agenda item default to participants of the meeting
GN-1959
start out without any participants on treatments and have the modal preset with
the participants of the meeting. This allows us to present the latest state of
the participants of the meeting and a user only has to click save to persist
this state.

Key differences:

OLD: treatments could have a different set on participants if the participants
of the meeting where changed between creating two agendapoints.

NEW: the user will have to fill in participants for every treatment
NEW: the user is presented with the participants of the meeting when setting the
participants of a treatment for the first time