Replace usage of hasDirtyAttributes
I mistakenly assumed it would also be set to true when a relationship
reference would be overwritten. This is not the case. While
the code still technically worked, it felt way too brittle to rely
on the fact that along with any relationship changes, this particular
file also always updated a property.