Make delete and update tasks unlinked
It was pretty easy to get the db in an inconsistent state
by cancelling the save/delete tasks halfway through. (for example
by clicking outside the modal while the tasks were running).

Because save and delete happen in steps (because of lacking support
for cascading in ED) this could cause problems since child
tasks are linked by default.