Added date object type check
With an instance check for Date, the helper no longer tries to format
things that are not dates, such as undefined or other objects.
In some instances, a date is left undefined, such as the end date on
meetings if the meeting isn't finished yet. We don't want unnecessary
errors being printend on the console because of this.