Fixed off by one errors on adding agendapoints
Putting the agendapoints before or after another agendapoint is already
usefull if there is at least 1 agendapoint in the list. This is when
`count > 0`, not `count > 1`.
Putting an agendapoint at the end of the collection, means putting it on the
index of `list.length` and not `list.length - 1`, given that collections start
at index 0 in JavaScript.