Trim title on validation and on saving Agendapoint
Use trim to remove the whitespace around the text and see if the length
is still greater than 0. Also trim the title when saving, because extra
spaces are sometimes shown (when editing the title while editing the
agendapoint) and sometimes not (in the inbox and while not editing the
title) due to the way HTML collapses whitespace, but not in inputs.
This hopefully reduces confusion.