fetch currentVersion in all agendapoint subroutes
a parent route is not reloaded when switching between child routes, since we
loaded currentVersion there this could become outdated after a save. This in
turn caused unexpected behaviour when switching between subroutes (e.g.
revisions and edit), where either fewer revisions were shown than actually
existed or an older version of the document was shown when returning to the edit route