Remove selection styling hacks that seem not relevant and broke links
This was introduced here: https://github.com/lblod/frontend-gelinkt-notuleren/pull/448
In order to tweak the styling of embedded regulatory statements in
agenda points. Since it no longer seems possible to select these PM
nodes, this styling never actually gets applied to these. The background
color was however being applied to any selection, not just those in
embedded regulatory statements, so was causing the input for the url of
a link to have an invisible selection.